+++ africanus/model/coherency/conversion.py
@@ -2,15 +2,21 @@
 
 
 from collections import OrderedDict, deque
+from dataclasses import dataclass
+from enum import Enum
+import heapq
 from pprint import pformat
 from textwrap import fill
+from typing import Any, Callable, Tuple
 
 import numpy as np
+import numpy.typing as npt
 
 from africanus.util.casa_types import STOKES_TYPES, STOKES_ID_MAP
 from africanus.util.docs import DocstringTemplate
 
-stokes_conv = {
+# Definitions for conversion from stokes to correlations
+STOKES_TO_CORR_CONV = {
     "RR": {("I", "V"): lambda i, v: i + v + 0j},
     "RL": {("Q", "U"): lambda q, u: q + u * 1j},
     "LR": {("Q", "U"): lambda q, u: q - u * 1j},
@@ -19,23 +25,60 @@ stokes_conv = {
     "XY": {("U", "V"): lambda u, v: u + v * 1j},
     "YX": {("U", "V"): lambda u, v: u - v * 1j},
     "YY": {("I", "Q"): lambda i, q: i - q + 0j},
+}
+
+# Definitions for conversion from correlations to stokes
+CORR_TO_STOKES_CONV = {
     "I": {
-        ("XX", "YY"): lambda xx, yy: (xx + yy).real / 2,
-        ("RR", "LL"): lambda rr, ll: (rr + ll).real / 2,
+        ("XX", "YY"): lambda xx, yy: (xx + yy) / 2,
+        ("RR", "LL"): lambda rr, ll: (rr + ll) / 2,
     },
     "Q": {
-        ("XX", "YY"): lambda xx, yy: (xx - yy).real / 2,
-        ("RL", "LR"): lambda rl, lr: (rl + lr).real / 2,
+        ("XX", "YY"): lambda xx, yy: (xx - yy) / 2,
+        ("RL", "LR"): lambda rl, lr: (rl + lr) / 2,
     },
     "U": {
-        ("XY", "YX"): lambda xy, yx: (xy + yx).real / 2,
-        ("RL", "LR"): lambda rl, lr: (rl - lr).imag / 2,
+        ("XY", "YX"): lambda xy, yx: (xy + yx) / 2,
+        ("RL", "LR"): lambda rl, lr: (rl - lr) / 2j,
     },
     "V": {
-        ("XY", "YX"): lambda xy, yx: (xy - yx).imag / 2,
-        ("RR", "LL"): lambda rr, ll: (rr - ll).real / 2,
+        ("XY", "YX"): lambda xy, yx: (xy - yx) / 2j,
+        ("RR", "LL"): lambda rr, ll: (rr - ll) / 2,
     },
 }
+
+
+CONVERSION_SCHEMA = {**STOKES_TO_CORR_CONV, **CORR_TO_STOKES_CONV}
+
+DataSource = Enum("DataSource", ["Index", "Default"])
+
+
+@dataclass(slots=True)
+class ProductMapping:
+    """Defines a mapping between two datasources and a destination via a callable function
+
+    Implements partial ordering such that mapping from actual data (DataSource.Index)
+    are preferred to DataSources producing default values (DataSource.Default)
+    """
+
+    source_one: Tuple[DataSource, Any]
+    source_two: Tuple[DataSource, Any]
+    dest_index: Tuple[Any, ...]
+    fn: Callable
+    dtype: npt.DTypeLike
+
+    @property
+    def priority(self):
+        """Prefer Index Data sources to Default Data sources"""
+        return int(self.source_one[0] == DataSource.Index) + int(
+            self.source_two[0] == DataSource.Index
+        )
+
+    def __lt__(self, other):
+        if not isinstance(other, ProductMapping):
+            raise NotImplementedError
+        # The priority reversal is intentional
+        return self.priority > other.priority
 
 
 class DimensionMismatch(Exception):
@@ -65,8 +108,7 @@ def _element_indices_and_shape(data):
             shape.append(len(current))
         elif shape[depth] != len(current):
             raise DimensionMismatch(
-                "Dimension mismatch %d != %d at depth %d"
-                % (shape[depth], len(current), depth)
+                f"Dimension mismatch {shape[depth]} != {len(current)} at depth {depth}"
             )
 
         # Handle each sequence element
@@ -77,7 +119,7 @@ def _element_indices_and_shape(data):
             # String
             elif isinstance(e, str):
                 if e in result:
-                    raise ValueError("'%s' defined multiple times" % e)
+                    raise ValueError(f"'{e}' defined multiple times")
 
                 result[e] = current_idx + (i,)
             # We have a CASA integer Stokes ID, convert to string
@@ -86,21 +128,20 @@ def _element_indices_and_shape(data):
                     e = STOKES_ID_MAP[e]
                 except KeyError:
                     raise ValueError(
-                        "Invalid id '%d'. "
-                        "Valid id's '%s'" % (e, pformat(STOKES_ID_MAP))
+                        f"Invalid id '{e}'. Valid id's '{pformat(STOKES_ID_MAP)}'"
                     )
 
                 if e in result:
-                    raise ValueError("'%s' defined multiple times" % e)
+                    raise ValueError(f"'{e}' defined multiple times")
 
                 result[e] = current_idx + (i,)
             else:
-                raise TypeError("Invalid type '%s' for element '%s'" % (type(e), e))
+                raise TypeError(f"Invalid type '{type(e)}' for element '{e}'")
 
     return result, tuple(shape)
 
 
-def convert_setup(input, input_schema, output_schema):
+def convert_setup(input, input_schema, output_schema, implicit_stokes):
     input_indices, input_shape = _element_indices_and_shape(input_schema)
     output_indices, output_shape = _element_indices_and_shape(output_schema)
 
@@ -113,45 +154,53 @@ def convert_setup(input, input_schema, output_schema):
     # Figure out how to produce an output from available inputs
     for okey, out_idx in output_indices.items():
         try:
-            deps = stokes_conv[okey]
+            deps = CONVERSION_SCHEMA[okey]
         except KeyError:
             raise ValueError(
-                "Unknown output '%s'. Known types '%s'" % (deps, STOKES_TYPES)
+                f"Unknown output {okey}. Known outputs: {list(CONVERSION_SCHEMA.keys())}"
             )
 
-        found_conv = False
+        # We can substitute defaults for stokes values when converting to correlations
+        # This makes it possible to compute mappings such as
+        # ['I'] -> ['XX', 'XY', 'YX', 'YY']
+        can_substitute_defaults = implicit_stokes and okey in STOKES_TO_CORR_CONV
+        okey_mappings = []
 
-        # Find a mapping for which we have inputs
+        # Find a mapping for which we have inputs or defaults
         for (c1, c2), fn in deps.items():
-            # Get indices for both correlations
-            try:
-                c1_idx = (Ellipsis,) + input_indices[c1]
-            except KeyError:
+            if c1 in input_indices:
+                c1_src = (DataSource.Index, (Ellipsis,) + input_indices[c1])
+            elif can_substitute_defaults:
+                c1_src = (DataSource.Default, 0)
+            else:
                 continue
 
-            try:
-                c2_idx = (Ellipsis,) + input_indices[c2]
-            except KeyError:
+            if c2 in input_indices:
+                c2_src = (DataSource.Index, (Ellipsis,) + input_indices[c2])
+            elif can_substitute_defaults:
+                c2_src = (DataSource.Default, 0)
+            else:
                 continue
 
-            found_conv = True
             out_idx = (Ellipsis,) + out_idx
             # Figure out the data type for this output
             dtype = fn(dummy, dummy).dtype
-            mapping.append((c1_idx, c2_idx, out_idx, fn, dtype))
-            break
-
-        # We must find a conversion
-        if not found_conv:
-            raise MissingConversionInputs(
-                "None of the supplied inputs '%s' "
-                "can produce output '%s'. It can be "
-                "produced by the following "
-                "combinations '%s'." % (input_schema, okey, deps.keys())
+            heapq.heappush(
+                okey_mappings, ProductMapping(c1_src, c2_src, out_idx, fn, dtype)
             )
 
-    out_dtype = np.result_type(*[dt for _, _, _, _, dt in mapping])
+        if len(okey_mappings) == 0:
+            raise MissingConversionInputs(
+                f"None of the supplied inputs '{input_schema}' "
+                f"can produce output '{okey}'. It can be "
+                f"produced by the following "
+                f"combinations '{deps.keys()}'."
+            )
 
+        # Use the highest priority mapping
+        mapping.append(okey_mappings[0])
+
+    out_dtype = np.result_type(*[m.dtype for m in mapping])
     return mapping, input_shape, output_shape, out_dtype
 
 
@@ -160,18 +209,22 @@ def convert_impl(input, mapping, in_shape, out_shape, dtype):
     out_shape = input.shape[: -len(in_shape)] + out_shape
     output = np.empty(out_shape, dtype=dtype)
 
-    for c1_idx, c2_idx, out_idx, fn, _ in mapping:
-        output[out_idx] = fn(input[c1_idx], input[c2_idx])
+    for m in mapping:
+        c1_type, c1_val = m.source_one
+        c2_type, c2_val = m.source_two
+        c1_arg = c1_val if c1_type is DataSource.Default else input[c1_val]
+        c2_arg = c2_val if c2_type is DataSource.Default else input[c2_val]
+        output[m.dest_index] = m.fn(c1_arg, c2_arg)
 
     return output
 
 
-def convert(input, input_schema, output_schema):
+def convert(input, input_schema, output_schema, implicit_stokes=False):
     """See STOKES_DOCS below"""
 
     # Do the conversion
     mapping, in_shape, out_shape, dtype = convert_setup(
-        input, input_schema, output_schema
+        input, input_schema, output_schema, implicit_stokes
     )
 
     return convert_impl(input, mapping, in_shape, out_shape, dtype)
@@ -238,6 +291,9 @@ input_schema : list of str or int
 output_schema : list of str or int
     A schema describing the :code:`ocorr_1, ..., ocorr_n`
     dimension of the return value.
+implicit_stokes : bool
+    A flag controlling whether implicitly assuming zeros
+    for missing Stokes inputs is allowed. Defaults to False.
 
 Returns
 -------